requirements.txt
alembic==1.0.6
branca==0.4.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
MarkupSafe==1.1.1
Flask-Login==0.5.0
Flask-Misaka==1.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.3
gunicorn==19.9.0
idna==2.8
Jinja2==2.10.1
Mako==1.0.7
misaka==2.1.1
python-dateutil==2.8.1
python-dotenv==0.10.1
requests==2.21.0
retrying==1.3.3
six==1.12.0
SQLAlchemy==1.3.0
Werkzeug==1.0.1
WTForms==2.3.3
